==> feature/0_data_sort_out.ipynb <==
import pandas as pd
import numpy as np
import datetime
import time

from tools.pos_encoder import *
from tools.time_encoder import *

from IPython.display import display, HTML
def show_df(df):
    display(HTML(df.to_html()))

# Read

train_dataset_path = '../data/eleme_round1_train/'
test_dataset_path = '../data/eleme_round1_testA/'
testB_dataset_path = '../data/eleme_round1_testB/'

def read_a_day(dataset_path, read_date):
    df_action = pd.read_csv('%saction/action_%s.txt' % (dataset_path, read_date.strftime('%Y%m%d')))
    df_courier = pd.read_csv('%scourier/courier_%s.txt' % (dataset_path, read_date.strftime('%Y%m%d')))
    df_distance = pd.read_csv('%sdistance/distance_%s.txt' % (dataset_path, read_date.strftime('%Y%m%d')))
    df_order = pd.read_csv('%sorder/order_%s.txt' % (dataset_path, read_date.strftime('%Y%m%d')))
    return df_action, df_courier, df_distance, df_order

def read_days(start_date, end_date, data_dir):
    cur_date = start_date
    actions = []
    couriers = []
    distances = []
    orders = []
    while cur_date < end_date:
        df_action, df_courier, df_distance, df_order = read_a_day(data_dir, cur_date)
        df_action['date'] = cur_date
        df_courier['date'] = cur_date
        df_distance['date'] = cur_date
        df_order['date'] = cur_date
        
        actions.append(df_action)
        couriers.append(df_courier)
        distances.append(df_distance)
        orders.append(df_order)
        cur_date += datetime.timedelta(days = 1)
    
    df_actions = pd.concat(actions, axis = 0, ignore_index = True)
    df_couriers = pd.concat(couriers, axis = 0, ignore_index = True)    
    df_distances = pd.concat(distances, axis = 0, ignore_index = True)    
    df_orders = pd.concat(orders, axis = 0, ignore_index = True) 
    return df_actions, df_couriers, df_distances, df_orders

df_actions_train, df_couriers_train, df_distances_train, df_orders_train = read_days(datetime.datetime(2020, 2, 1), datetime.datetime(2020, 3, 1), train_dataset_path)
df_actions_testA, df_couriers_testA, df_distances_testA, df_orders_testA = read_days(datetime.datetime(2020, 3, 1), datetime.datetime(2020, 3, 7), test_dataset_path)

df_actions_testB, df_couriers_testB, df_distances_testB, df_orders_testB = read_days(datetime.datetime(2020, 3, 1), datetime.datetime(2020, 3, 7), testB_dataset_path)

df_orders_testA

# build dataset-level features

## one hot

def one_hot_features(df_train, df_test, one_hot_features):
    df_train['type'] = 'train'
    df_test['type'] = 'test'
    one_hot_names = []
    df = pd.concat([df_train, df_test])
    for feature in one_hot_features:
        df_one_hot = pd.get_dummies(df[feature], prefix=feature)
        df = pd.concat([df, df_one_hot], axis=1)
        one_hot_names.extend(list(df_one_hot.columns))
    return df.query('type == "train"'), df.query('type == "test"'), one_hot_names

df_actions_train, df_actions_testA, actions_one_hot_names = one_hot_features(df_actions_train, df_actions_testA, ['action_type'])
df_couriers_train, df_couriers_testA, couriers_one_hot_names = one_hot_features(df_couriers_train, df_couriers_testA, ['level', 'max_load'])
df_distances_train, df_distances_testA, distances_one_hot_names = one_hot_features(df_distances_train, df_distances_testA, ['source_type', 'target_type'])
df_orders_train, df_orders_testA, orders_one_hot_names = one_hot_features(df_orders_train, df_orders_testA, ['weather_grade'])

df_couriers_train['level'].value_counts()

## order aoi shop courier 交叉

df_order = pd.concat([df_orders_train, df_orders_testA])

def build_aoi_feature(df):
    res_dict = {'aoi_count' : df.shape[0]}
    res_dict['aoi_shop_count'] = len(set(df['shop_id']))
    res_dict['aoi_courier_count'] = len(set(df['courier_id']))
    return pd.Series(res_dict)

def build_shop_feature(df):
    res_dict = {'shop_count' : df.shape[0]}
    res_dict['shop_aoi_count'] = len(set(df['aoi_id']))
    res_dict['shop_courier_count'] = len(set(df['courier_id']))
    return pd.Series(res_dict)

df_feature_aoi = df_order.groupby('aoi_id').apply(build_aoi_feature)
df_feature_shop = df_order.groupby('shop_id').apply(build_shop_feature)
df_feature_courier = df_order.groupby('courier_id').apply(lambda x : pd.Series({
    'courier_aoi_count' : len(set(x['aoi_id'])),
    'courier_shop_count' : len(set(x['shop_id'])),
    'courier_count' : x.shape[0]
}))

def add_aoi_shop_features(df_origin, df_feature, origin_name):
    feature_names = df_feature.columns
    for feature in feature_names:
        df_origin[feature] = df_origin[origin_name].map(df_feature[feature])
    return df_origin


df_order = add_aoi_shop_features(df_order, df_feature_aoi, 'aoi_id')
df_order = add_aoi_shop_features(df_order, df_feature_shop, 'shop_id')
df_order = add_aoi_shop_features(df_order, df_feature_courier, 'courier_id')

df_orders_train, df_orders_testA = df_order.query('type == "train"'), df_order.query('type == "test"')

# Map

def get_map(df):
    """
    将数据映射成字典，可通过
        mp[date][courier_id][wave_index]
    访问

    """
    last_date, last_courier, last_wave = df.at[0, 'date'], df.at[0, 'courier_id'], df.at[0, 'wave_index']
    last_i = 0
    mp_date, mp_courier, mp_wave = {}, {}, {}
    key_list = []
    for i in range(1, df.shape[0]):
        cur_wave = df.at[i, 'wave_index']
        cur_courier = df.at[i, 'courier_id']
        cur_date = df.at[i, 'date']

        if last_wave != cur_wave or (last_wave == cur_wave and last_courier != cur_courier):
            mp_wave[last_wave] = df.iloc[last_i : i]
            last_i = i
            key_list.append([last_date, last_courier, last_wave])
            last_wave = cur_wave

        if last_courier != cur_courier or (last_courier == cur_courier and last_date != cur_date):
            mp_courier[last_courier] = mp_wave.copy()
            mp_wave = {}
            last_courier = cur_courier

        if last_date != cur_date:
            mp_date[last_date] = mp_courier.copy()
            mp_courier = {}
            last_date = cur_date
    
    mp_wave[last_wave] = df.iloc[last_i : ]
    key_list.append([last_date, last_courier, last_wave])
    mp_courier[last_courier] = mp_wave.copy()
    mp_date[last_date] = mp_courier.copy()
    mp_courier = {}
    
    return mp_date, key_list


start = time.time()
mp_action_train, action_key_list_train = get_map(df_actions_train)
mp_distance_train, distance_key_list_train = get_map(df_distances_train)
mp_order_train, order_key_list_train = get_map(df_orders_train)
time.time() - start

start = time.time()
mp_action_testA, action_key_list_testA = get_map(df_actions_testA)
mp_distance_testA, distance_key_list_testA = get_map(df_distances_testA)
mp_order_testA, order_key_list_testA = get_map(df_orders_testA)
time.time() - start

start = time.time()
mp_action_testB, action_key_list_testB = get_map(df_actions_testB)
mp_distance_testB, distance_key_list_testB = get_map(df_distances_testB)
mp_order_testB, order_key_list_testB = get_map(df_orders_testB)
time.time() - start

def get_map_distance_detail(mp_distance, key_list):
    mp_distance_detail = {}
    for (date, courier, wave_idx) in key_list:
        if date not in mp_distance_detail:
            mp_distance_detail[date] = {}
        if courier not in mp_distance_detail[date]:
            mp_distance_detail[date][courier] = {}
        if wave_idx not in mp_distance_detail[date][courier]:
            mp_distance_detail[date][courier][wave_idx] = {}
        
        df_a_distance = mp_distance[date][courier][wave_idx]
        
        gby_tracking = df_a_distance.groupby('tracking_id')
        for tracking_id, df_track_id in gby_tracking:
            mp_distance_detail[date][courier][wave_idx][tracking_id] = {}
            gby_target_tracking = df_track_id.groupby('target_tracking_id')
            for target_trackingid, df_target_tracking in gby_target_tracking:
                mp_distance_detail[date][courier][wave_idx][tracking_id][target_trackingid] = df_target_tracking
            
    return mp_distance_detail

start = time.time()
mp_distance_detail_train = get_map_distance_detail(mp_distance_train, action_key_list_train)
mp_distance_detail_testA = get_map_distance_detail(mp_distance_testA, action_key_list_testA)

time.time() - start

mp_distance_detail_testB = get_map_distance_detail(mp_distance_testB, action_key_list_testB)


# Check

len(action_key_list_train), len(distance_key_list_train), len(order_key_list_train), \
len(action_key_list_testA), len(distance_key_list_testA), len(order_key_list_testA)

len(action_key_list_testB), len(distance_key_list_testB), len(order_key_list_testB)

sorted(action_key_list_train) == sorted(distance_key_list_train) == sorted(order_key_list_train), \
sorted(action_key_list_testA) == sorted(distance_key_list_testA) == sorted(order_key_list_testA)

sorted(action_key_list_testB) == sorted(distance_key_list_testB) == sorted(order_key_list_testB)

date = datetime.datetime(2020, 3, 6)
df_actions_testA.query('date == @date & courier_id == "10325553"') 

date = datetime.datetime(2020, 3, 6)
mp_action_testA[date][10325553]

i = 0
for (a, b, c), (aa, bb, cc) in zip(sorted(action_key_list_testA), sorted(distance_key_list_testA)):
    if a != aa or b != bb or c != cc:
        print(i)
    i += 1
        

# pickle

import pickle
PICKLE_PATH = '../user_data/data_sort_out'

testB_dataset_mp = {
    'key_list' : action_key_list_testB,
    'mp_action' : mp_action_testB,
    'mp_distance' : mp_distance_testB,
    'mp_order' : mp_order_testB,
    'mp_distance_detail' : mp_distance_detail_testB
}
with open('%s/testB_dataset_mp.pickle'%(PICKLE_PATH), 'wb') as f:
    pickle.dump(testB_dataset_mp, f)

testB_dataset_df = {
    'df_actions' : df_actions_testB,
    'df_couriers' : df_couriers_testB,
    'df_distances' : df_distances_testB,
    'df_orders' : df_orders_testB
}
with open('%s/testB_dataset_df.pickle'%(PICKLE_PATH), 'wb') as f:
    pickle.dump(testB_dataset_df, f)

train_dataset_mp = {
    'key_list' : action_key_list_train,
    'mp_action' : mp_action_train,
    'mp_distance' : mp_distance_train,
    'mp_order' : mp_order_train,
    'mp_distance_detail' : mp_distance_detail_train
}
with open('%s/train_dataset_mp.pickle'%(PICKLE_PATH), 'wb') as f:
    pickle.dump(train_dataset_mp, f)

testA_dataset_mp = {
    'key_list' : action_key_list_testA,
    'mp_action' : mp_action_testA,
    'mp_distance' : mp_distance_testA,
    'mp_order' : mp_order_testA,
    'mp_distance_detail' : mp_distance_detail_testA
}
with open('%s/testA_dataset_mp.pickle'%(PICKLE_PATH), 'wb') as f:
    pickle.dump(testA_dataset_mp, f)

train_dataset_df = {
    'df_actions' : df_actions_train,
    'df_couriers' : df_couriers_train,
    'df_distances' : df_distances_train,
    'df_orders' : df_orders_train
}
with open('%s/train_dataset_df.pickle'%(PICKLE_PATH), 'wb') as f:
    pickle.dump(train_dataset_df, f)

testA_dataset_df = {
    'df_actions' : df_actions_testA,
    'df_couriers' : df_couriers_testA,
    'df_distances' : df_distances_testA,
    'df_orders' : df_orders_testA
}
with open('%s/testA_dataset_df.pickle'%(PICKLE_PATH), 'wb') as f:
    pickle.dump(testA_dataset_df, f)

one_hot_name = {
    'actions' : actions_one_hot_names,
    'couriers' : couriers_one_hot_names,
    'distances' : distances_one_hot_names,
    'orders' : orders_one_hot_names
}
with open('%s/one_hot_name.pickle'%(PICKLE_PATH), 'wb') as f:
    pickle.dump(one_hot_name, f)
